requirements.txt
ipython
jupyter
numpy
scipy
pandas
matplotlib
seaborn
ruamel.yaml
pyyaml
tqdm
Pillow
